feat: Validations in UserService and ProcessService

+++ src/main/java/com/limatech/juriprocessos/services/process/ProcessService.java
@@ -2,18 +2,22 @@ package com.limatech.juriprocessos.services.process;
 
 import com.limatech.juriprocessos.dtos.process.CreateProcessDTO;
 import com.limatech.juriprocessos.exceptions.process.ProcessNotFoundException;
+import com.limatech.juriprocessos.exceptions.users.ForbiddenActionException;
 import com.limatech.juriprocessos.exceptions.users.UserNotFoundException;
 import com.limatech.juriprocessos.models.process.entity.Process;
 import com.limatech.juriprocessos.models.users.entity.User;
 import com.limatech.juriprocessos.repository.process.ProcessRepository;
 import com.limatech.juriprocessos.repository.users.UserRepository;
+import com.limatech.juriprocessos.services.interfaces.UserValidation;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.UUID;
 
 @Service
-public class ProcessService {
+public class ProcessService implements UserValidation {
 
     private final ProcessRepository processRepository;
     private final UserRepository userRepository;
@@ -25,6 +29,8 @@ public class ProcessService {
     }
 
     public Process createProcess(CreateProcessDTO processDTO) {
+        validateUserPermission(processDTO);
+
         User user = userRepository.findById(processDTO.getUserId()).orElseThrow(() -> new UserNotFoundException("User not found"));
         processDTO.setUser(user);
         Process process = processDTO.toEntity();
@@ -33,12 +39,16 @@ public class ProcessService {
     }
 
     public void deleteProcess(UUID id) {
+        validateUserPermission(id);
+
         Process process = processRepository.findById(id).orElseThrow(() -> new ProcessNotFoundException("Process not found"));
 
         processRepository.deleteById(id);
     }
 
     public Process updateProcess(UUID id, CreateProcessDTO processDTO) {
+        validateUserPermission(id);
+
         Process process = processRepository.findById(id).orElseThrow(() -> new ProcessNotFoundException("Process not found"));
 
         if(processDTO.getIdentifier() != null) {
@@ -61,6 +71,38 @@ public class ProcessService {
     }
 
     public Process getProcess(UUID id) {
+        validateUserPermission(id);
+
         return processRepository.findById(id).orElseThrow(() -> new ProcessNotFoundException("Process not found"));
+    }
+
+    @Override
+    public void validateUserPermission(UUID id) {
+        User currentUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        UUID currentUserId = currentUser.getId();
+
+        User currentUserWithProcess = userRepository.findById(currentUserId).orElseThrow(() -> new UserNotFoundException("Current user not " +
+                        "found"));
+
+        List<Process> processes = currentUserWithProcess.getProcesses();
+        List<UUID> processesID = processes.stream().map(Process::getId).toList();
+
+        System.out.println(processesID.size());
+        if(!processesID.contains(id) && !isUserAdmin()) {
+            throw new ForbiddenActionException();
+        }
+    }
+
+    public void validateUserPermission(CreateProcessDTO processDTO) {
+        UUID userFromProcess = processDTO.getUserId();
+
+        User currentUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        UUID currentUserId = currentUser.getId();
+
+
+        if(!userFromProcess.toString().equals(currentUserId.toString()) && !isUserAdmin()) {
+            throw new ForbiddenActionException();
+        }
+
     }
 }
